+++ src/main/java/org/ramer/diary/controller/user/FavouriteController.java
@@ -3,19 +3,20 @@
  */
 package org.ramer.diary.controller.user;
 
-import java.io.IOException;
-
-import javax.annotation.Resource;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-
 import org.ramer.diary.domain.Topic;
 import org.ramer.diary.domain.User;
 import org.ramer.diary.service.FavouriteService;
 import org.ramer.diary.util.UserUtils;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.SessionAttributes;
+
+import javax.annotation.Resource;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
+import java.io.IOException;
 
 /**
  * 收藏和取消收藏.
@@ -24,7 +25,7 @@ import org.springframework.web.bind.annotation.*;
  */
 @SessionAttributes(value = { "user", "topics", }, types = { User.class, Topic.class })
 @Controller
-public class AddFavourite{
+public class FavouriteController{
     @Resource
     private FavouriteService favouriteService;
 
@@ -43,10 +44,6 @@ public class AddFavourite{
             HttpSession session) throws IOException {
         response.setCharacterEncoding("UTF-8");
         if (!UserUtils.checkLogin(session)) {
-            //            if (!UserUtils.multiLogin(session, u)) {
-            //                response.getWriter().write("账号异地登陆！ 当前登陆失效，如果不是本人，请及时修改密码 !");
-            //                return;
-            //            }
             response.getWriter().write("麻麻说没登录不能收藏哒 !");
             return;
         }
