--- src/main/java/org/ramer/diary/controller/user/Login.java
@@ -1,26 +1,0 @@
-package org.ramer.diary.controller.user;
-
-import lombok.extern.slf4j.Slf4j;
-import org.ramer.diary.domain.Topic;
-import org.ramer.diary.domain.User;
-import org.ramer.diary.service.UserService;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.SessionAttributes;
-
-import javax.annotation.Resource;
-
-/**
- * 登陆类.
- * @author ramer
- *
- */
-@Slf4j
-@SessionAttributes(value = { "user", "topics", }, types = { User.class, Topic.class })
-@Controller
-public class Login{
-    @Resource
-    private UserService userService;
-
-
-
-}
