+++ src/main/java/org/ramer/diary/controller/user/PraiseController.java
@@ -1,23 +1,23 @@
 package org.ramer.diary.controller.user;
 
-import java.io.IOException;
-import java.util.Map;
+import lombok.extern.slf4j.Slf4j;
+import org.ramer.diary.domain.Topic;
+import org.ramer.diary.domain.User;
+import org.ramer.diary.exception.DiaryException;
+import org.ramer.diary.service.PraiseService;
+import org.ramer.diary.service.UserService;
+import org.ramer.diary.util.UserUtils;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.SessionAttributes;
 
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
-
-import org.ramer.diary.domain.Topic;
-import org.ramer.diary.domain.User;
-import org.ramer.diary.exception.UserNotLoginException;
-import org.ramer.diary.service.PraiseService;
-import org.ramer.diary.service.UserService;
-import org.ramer.diary.util.UserUtils;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.*;
-
-import lombok.extern.slf4j.Slf4j;
+import java.io.IOException;
+import java.util.Map;
 
 /**
  * 点赞和取消点赞.
@@ -27,7 +27,7 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 @SessionAttributes(value = { "user", "topics", }, types = { User.class, Topic.class })
 @Controller
-public class AddPraise{
+public class PraiseController{
 
     @Resource
     private PraiseService praiseService;
@@ -84,7 +84,7 @@ public class AddPraise{
         response.setCharacterEncoding("utf-8");
         if (!UserUtils.checkLogin(session)) {
             User u = userService.getById(((User) session.getAttribute("user")).getId());
-            throw new UserNotLoginException("没登录的哦");
+            throw new DiaryException("没登录的哦");
         }
         Topic topic = new Topic();
         topic.setId(topic_id);
