--- src/main/java/org/ramer/diary/exception/PasswordNotMatchException.java
@@ -1,21 +1,0 @@
-/*
- *
- */
-
-package org.ramer.diary.exception;
-
-/**
- * 密码不匹配异常处理
- * @author ramer
- *
- */
-public class PasswordNotMatchException extends RuntimeException {
-  private static final long serialVersionUID = 1L;
-
-  public PasswordNotMatchException(String message) {
-    super(message);
-  }
-
-  public PasswordNotMatchException() {
-  }
-}
